resetting the player
Hello, I edited the CountDownTimer script so that it resets the player to the start of the level when the timer times out. This is done by teleporting the player to an invisible checkpoint at the very start of the level called StartPointCheckpoint. Technically it teleports the player 1 second before the timer times out.